调整：使用ItemStackHandler来处理容器存储

+++ src/main/java/mczme/lingshi/common/datamap/ingredient/CookingFoodData.java
@@ -4,10 +4,10 @@ import com.mojang.serialization.Codec;
 import com.mojang.serialization.codecs.RecordCodecBuilder;
 
 
-public record CookingFoodData(float cookedTime,float burntTime) {
-    static float MAX_TIME = 30;
+public record CookingFoodData(int cookedTime,int burntTime) {
+    static int MAX_TIME = 30;
 
-    public CookingFoodData(float cookedTime,float burntTime){
+    public CookingFoodData(int cookedTime,int burntTime){
         if(cookedTime > MAX_TIME) this.cookedTime = MAX_TIME;
         else if(cookedTime < 0) this.cookedTime = 0;
         else this.cookedTime = cookedTime;
@@ -15,8 +15,8 @@ public record CookingFoodData(float cookedTime,float burntTime) {
     }
 
     public static final Codec<CookingFoodData> CODEC = RecordCodecBuilder.create(instance -> instance.group(
-            Codec.FLOAT.fieldOf("cooked_time").forGetter(CookingFoodData::cookedTime),
-            Codec.FLOAT.fieldOf("complete_time").forGetter(CookingFoodData::burntTime)
+            Codec.INT.fieldOf("cooked_time").forGetter(CookingFoodData::cookedTime),
+            Codec.INT.fieldOf("complete_time").forGetter(CookingFoodData::burntTime)
     ).apply(instance, CookingFoodData::new));
 
 }
